migrate location feature
